upgrade to react native 0.36.0
